# frontend/admin/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RayFlow API",
    "description": "🐟 RayFlow - Flink/Fluss/Paimon 任务发布管理平台 API 文档",
    "contact": {
      "name": "RayFlow Team"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.0.4"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/image-registry": {
      "get": {
        "tags": [
          "Image Registry Management"
        ],
        "summary": "获取当前组织镜像仓库配置",
        "operationId": "getImageRegistryConfig",
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImageRegistryConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Image Registry Management"
        ],
        "summary": "保存当前组织镜像仓库配置",
        "operationId": "updateImageRegistryConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRegistryConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImageRegistryConfigResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/image-registry/test": {
      "post": {
        "tags": [
          "Image Registry Management"
        ],
        "summary": "测试镜像仓库配置",
        "operationId": "testImageRegistryConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRegistryConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImageRegistryConfigRequest": {
        "required": [
          "registryUrl"
        ],
        "type": "object",
        "properties": {
          "registryUrl": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string"
          },
          "namespaceName": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "username": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "ImageRegistryConfigResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "registryUrl": {
            "type": "string"
          },
          "namespaceName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}